requirements.txt
fastapi
uvicorn
python-docx
requests
jinja2
azure-identity
azure-keyvault-secrets
bibtexparser
